Pulled cordova app object out of main.js
